use fail point to stabilize the new test
